requirements.txt
lxml==4.2.1
scipy==0.19.1
numpy==1.13.3
Keras==2.2.4
nltk==3.2.5
dataset==1.1.0
python_dateutil==2.7.5
scikit_learn==0.20.1
dill=0.2.8.2
